+++ tests/testthat/test-download.R
@@ -1,5 +1,6 @@
+random <- paste(sample(letters, 16), collapse = "")
+
 test_that("Downloading without preper credentials raises an error", {
-  random <- sample(letters, 8)
 
   # arrange
   r_env_file <- construct_Renviron_path()
@@ -14,8 +15,22 @@ test_that("Downloading without preper credentials raises an error", {
   }
 
   # act & assert
-  expect_error(download(construct_url(717)))
+  expect_error(download(construct_url(717)), regexp = "Check your credentials")
 
   # clean up
   file.copy(r_env_backup, r_env_file, overwrite = TRUE)
+})
+
+test_that("Downloader refuses unexpected URLs", {
+  error_regex <- "refus\\w{,3} to connect"
+
+  expect_error(
+    download(paste0("http://evil.", random, "-api.net/?format=json")),
+    error_regex
+  )
+
+  expect_error(
+    download(paste0("https://bacdiveZdsmz.de/api/bacdive/?format=json")),
+    error_regex
+  )
 })